feat: OrderDetail ErrorBoundary 적용
