+++ src/components/ErrorBox.tsx
@@ -1,5 +1,7 @@
 import { styled } from "styled-components";
 import { ERROR_MESSAGE } from "../constants/index";
+import { Button } from "./Button";
+import { ROUTER_PATH } from "../router";
 
 interface ErrorType {
   status: keyof typeof ERROR_MESSAGE;
@@ -10,6 +12,9 @@ const ErrorBox = ({ status }: ErrorType) => {
     <Wrapper>
       <h2>{status}</h2>
       <p>{ERROR_MESSAGE[status]}</p>
+      <Button>
+        <a href={ROUTER_PATH.Main}>홈으로</a>
+      </Button>
     </Wrapper>
   );
 };
